==> Covid/1_CovidCase_getdata.Rmd <==
---
title: "Obtain data for the case study on Covid-19 B-cells"
author: "Jeroen Gilis & Laura Perin"
date: "14/11/2023"
output:
    html_document:
      code_download: true    
      theme: cosmo
      toc: true
      toc_float: true
      highlight: tango
      number_sections: true
---

# Load libraries

```{r, message=FALSE, warning=FALSE}
library(CuratedAtlasQueryR) #0.4.3 (or higher?)
library(SingleCellExperiment) #v1.20.1
library(Matrix)
library(here)
library(fs)
```

# Prepare storing objects

```{r}
here::i_am("1_CovidCase_getdata.Rmd")
out_dir <- here::here("objects")
fs::dir_create(out_dir)
```

# Get data

```{r}
metadata <- get_metadata()

sce <- metadata |>
    dplyr::filter(
        name == "Single-cell multi-omics analysis of the immune response in COVID-19",
        cell_type %in% c("B cell",
                         "naive B cell",
                         "immature B cell",
                         "unswitched memory B cell",
                         "class switched memory B cell"),
        disease %in% c("COVID-19",
                       "normal")
    ) |>
    get_single_cell_experiment()
colnames(colData(sce))[2] <- "cell_type_curated"
```

# Get author categories metadata

```{r}
data_id <- metadata |>
    dplyr::filter(
        name == "Single-cell multi-omics analysis of the immune response in COVID-19") |>
  pull("file_id") |>
  first()

unharmonised_meta <- get_unharmonised_metadata(data_id)
cd_authors <- dplyr::collect(unharmonised_meta[[data_id]])
cd_authors
```

Now include the relevant author coldata in the curated coldata. The former can
be matched to the `original_cell_id` column of the latter:

```{r}
relevant_author_vars <- c("total_counts", "pct_counts_mt", "initial_clustering",
                          "Resample", "Swab_result", "Status", "Smoker", 
                          "Status_on_day_collection", "Collection_Day",
                          "Status_on_day_collection_summary", "Days_from_onset",
                          "Site", "Worst_Clinical_Status", "Outcome", 
                          "donor_id", "author_cell_type")
cd_authors <- cd_authors[match(sce$original_cell_id, rownames(cd_authors)),
                         relevant_author_vars]
colData(sce) <- cbind(colData(sce),
                      cd_authors)
sce
```

# Filtering on author coldata

```{r}
table(sce$Status)
sce <- sce[,sce$Status %in% c("Covid", "Healthy")]
colData(sce) <- droplevels(colData(sce))
table(sce$Status)
```

# Unnormalize the counts

The abundance matrix that can be obtained through download with the 
`CuratedAtlasQueryR` package are not raw, integer counts. Instead, the 
expression values have been log-transformed and normalized for differences in
library size using `scanpy`. However, our downstream analyses (DD and DE 
workflows) require raw, integer counts. Therefore, we will here undo/revert the 
normalization by `scanpy` manually.

```{r}
# move away from H5AD backend and save as regular rds
counts(sce) <- as(counts(sce), "sparseMatrix")
```

```{r}
sce$total_counts <- cd_authors$total_counts[match(sce$original_cell_id,
                                                  cd_authors$cell_)]
# "unnormalize" the counts
raw <- counts(sce)
raw <- exp(raw) - 1
raw <- raw/1e4
raw <- sweep(raw, 2, sce$total_counts, "*")
raw <- round(raw) # not integer at ±8th decimal due to back-and-forth conversion
assays(sce)$counts <- as(raw, "sparseMatrix")
```

# Save Covid B cells object

```{r}
saveRDS(sce, "./objects/sce_Covid_Bcells.rds")
```

```{r}
sessionInfo()
```
